[updated] OR
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1642 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/scene/UpdateTransformMessage.h
@@ -67,7 +67,7 @@ public:
 		insertByte(0); // unknown
 
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte)scno->getDirectionAngle());
 	}
 
 	UpdateTransformMessage(SceneObject* scno, float posX, float posZ, float posY) : BaseMessage(50) {
